refactor pdffile model to data package
