+++ app/src/main/java/com/ferdinand/pdftestapp/mappers/PdfFileMapper.kt
@@ -1,6 +1,6 @@
 package com.ferdinand.pdftestapp.mappers
 
 import com.ferdinand.pdftestapp.data.models.DbFavoritePdfFile
-import com.ferdinand.pdftestapp.models.PdfFile
+import com.ferdinand.pdftestapp.data.PdfFile
 
 fun PdfFile.toDbModel() = DbFavoritePdfFile(this.id)
